Add files via upload
Hi, My name is Naveen and i am doing this MotionCut virtual internship.
This internship is based on web development and our first task is to develop an website that consists of and home page , contact page , products page and finally an about page. In this to develop the website i have used HTML, CSS, JAVASCRIPT, languages.
The website i have developed is an food delivery website.